requirements.txt
django==4.2
python-dotenv==1.1.0
psycopg2-binary==2.9.10
bson==0.5.10
djangorestframework==3.15.0